John Fucking Goblin
Added John Goblin with his own prefab that has a queue for each task. For right now the only functionality is for collecting. Fixed Plant Prefabs given the sprites and removed redundant prefabs. Added beginnings of dryout functionality. A coroutine in the plot that refers to each plot in it's region and tells the plant there to call the coroutine dryout (does nothing now, but will wait for an input later to be watered). The dryout coroutine on plot already checks itself for being a watery plot. Altered Touchcontroller so TileInfo includes the plant at the reference location for the dictionary to work better.